+++ st_unique_code/changelog.rst
@@ -2,7 +2,8 @@
 ------
 - **New:** added field product_seq_id. Creating 'ir.sequence' for each created partner
 - **New:** added field UPN(Unique Product Number). Generate automatically
-- **Improvement:** Changed the way to build a unique code for a field
+- **Improvement:** Notification when product created without partner
+- **Improvement:** Changed the way to build a unique code